--- a/cache/v2/cache.go
+++ b/cache/v2/cache.go
@@ -1,0 +1,11 @@
+package cache
+
+import "io"
+
+type Cache[K comparable, V any] interface {
+	Get(K) (V, bool, error)
+	Set(K, V) error
+	Evict(K) error
+
+	io.Closer
+}