--- a/cache/v2/policy/multi_policy_test.go
+++ b/cache/v2/policy/multi_policy_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestNopPolicy(t *testing.T) {
-	p := CombinePolicies()
+	p := CombinePolicies[string]()
 
 	select {
 	case <-p.C():