… cache/lock_cache.go
… cache/policy/size/policy.go
… cache/policy/time/policy.go
… cache/sharded_cache.go