Currently requires 1.24 because that implements `maphash.Comparable`. It can be backported if we want it sooner.

The old cache is implemented in terms of the new cache.

The goal of this v2 cache is to improve ergonomics of working with caches

The following benchmark shows the speed difference for the hash functions used, as we used to use fnv64a before but that was a bit of a pain to adapt to generic types. Instead, we use the stdlib maphash that should be more secure and about as fast for the majority of inputs.

If the speed difference is significant, we can implement a fallback where one would provide it's own hash function to the cache, with the v1 cache defaulting to fnv64a.

```
$ go test -bench=BenchmarkHash ./cache --count 10 | benchstat -col .name -
goos: linux
goarch: amd64
pkg: github.com/upfluence/pkg/cache
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
         │  HashFnv64   │             HashMaphash             │
         │    sec/op    │   sec/op     vs base                │
*/8-16      2.178µ ± 0%   4.167µ ± 1%  +91.34% (p=0.000 n=10)
*/16-16     3.788µ ± 1%   4.145µ ± 2%   +9.42% (p=0.000 n=10)
*/64-16    17.119µ ± 8%   4.611µ ± 3%  -73.06% (p=0.000 n=10)
*/256-16    52.51µ ± 4%   11.46µ ± 0%  -78.18% (p=0.000 n=10)
geomean     9.279µ        5.496µ       -40.77%
```

It does not take into account the overhead gained from unboxing the value
types, nor any possible PGO wins.

An example of the keys used are domains, in per-domain shop ratelimiters,
So I picked string sizes that may be representative of those.
(I forgot 32 and the bench runs for too long).

To put this change into context, I am currently in the process of figuring out more adequate rate-limiting in shopify (#VEL-4774 and the cache ergonomics looked not great since we have generics now). It looked like a quick win and nice QOL here we are.